[ui] Redesign homepage and add apps section
The home page has been modified to try to make clearer what the registry
is to the users. The middle section has been simplified, the search section
made more predominant and the number of publishers is shown.
The apps section contains a static template (not ckanext-apps), and two of
the screenshots need to be replaced.