[feeds] Show feed links in the dataset page
The general registry feed is always shown. When a search is performed
a link appears for subscribing to the results. There is a bit of logic
in the template (which should be in the controller) to link to the
country, publisher, etc. feeds directly if the search was limited to
them.