Allow to publish datasets on behalf of other organizations
This is stored as a metadata field in the dataset (secondary_publisher)
as a free text field. There is a new field on the form, a new facet on
the search form and the secondary publisher is shown on the dataset
page. Also the CSV import/export supports the new field.