Save issues during archive process
A bit of refactoring to spearate the archiving of a particular dataset.
If an issue is found (no URL, file missinh, wrong XML...). Three extras
are stored on the dataset, the issue type, a message and the date it was
created. On successive runs of the archiver if the type and message of
the issue have not changed, the dataset won't be updated. If the process
arrives to the point where the dataset should be updated, we assume that
the issue has been fixed and the extras are deleted.