Fix broken logic for Show more facet links
cc @johnmartin

The whole thing is pretty broken on core, but this template was even
more broken :). The last item of the facets was not found and the Show
more link did not appear. In order to make the case where there are
exactly 10 items, search.facets.default=11 is required in the ini file,
due to h.get_facet_items_dict ignoring the provided limit param.