Remove pending publishers from group_list
The logic in ckan/model/meta.py for setting a revision to current is
ignored when state = 'pending', which makes sense as a pending review is
not the definitive one. In our case the state should be 'pending_review'
or similar, but rather than changing the state value we will filter out
the pending in a custom group_list, which is bound to break less
things.